Null instance fix
I was getting this error when picking up items: "Attempt to call function 'is_in_group' in base 'null instance' on a null instance"

Video with the bug:
https://drive.google.com/file/d/12DoBIOTe3VrW-zxasPUBMvF7sPRaazza/view

The demo was working anyway, but it would take me out of the game. Adding the null check seems to fix it.